+++ Assets/LDtkUnity/Runtime/Data/Extensions/Definition/AutoLayerRuleDefinition.cs
@@ -8,16 +8,16 @@ namespace LDtkUnity
         /// <value>
         /// Pivot of a tile stamp (0-1)
         /// </value>
-        [IgnoreDataMember] public Vector2 UnityPivot => new Vector2((float)PivotX, (float)PivotY);
+        [IgnoreDataMember] public Vector2 UnityPivot => new Vector2(PivotX, PivotY);
         
         /// <value>
         /// Cell coord modulo
         /// </value>
-        [IgnoreDataMember] public Vector2Int UnityModulo => new Vector2Int((int)XModulo, (int)YModulo);
+        [IgnoreDataMember] public Vector2Int UnityModulo => new Vector2Int(XModulo, YModulo);
         
         /// <value>
         /// Cell start offset
         /// </value>
-        [IgnoreDataMember] public Vector2Int UnityOffset => new Vector2Int((int)XOffset, (int)YOffset);
+        [IgnoreDataMember] public Vector2Int UnityOffset => new Vector2Int(XOffset, YOffset);
     }
 }
