+++ Assets/LDtkUnity/Runtime/Data/Extensions/Instance/FieldInstanceGridPoint.cs
@@ -8,6 +8,6 @@ namespace LDtkUnity
         /// <value>
         /// Grid-based coordinate
         /// </value>
-        [IgnoreDataMember] public Vector2Int UnityCoord => new Vector2Int((int)Cx, (int)Cy);
+        [IgnoreDataMember] public Vector2Int UnityCoord => new Vector2Int(Cx, Cy);
     }
 }
